Modification of the Service, Mapper, DTO and Controller of Event and User.

+++ src/main/java/com/example/gastroValenciaApi/services/EventService.java
@@ -1,34 +1,45 @@
 package com.example.gastroValenciaApi.services;
 
+import com.example.gastroValenciaApi.dtos.EventDTO;
+import com.example.gastroValenciaApi.mappers.EventMapper;
 import com.example.gastroValenciaApi.models.EventModel;
 import com.example.gastroValenciaApi.repositories.EventRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class EventService {
 
-    private EventRepository eventRepository;
+    private final EventRepository eventRepository;
+    private final EventMapper eventMapper;
 
-    public EventService(EventRepository eventRepository) {
+    public EventService(EventRepository eventRepository, EventMapper eventMapper) {
         this.eventRepository = eventRepository;
+        this.eventMapper = eventMapper;
     }
 
-    public List<EventModel> getAllEvents() {
-        return eventRepository.findAll();
+    public List<EventDTO> getAllEvents() {
+        return eventRepository.findAll().stream().map(eventMapper::toDTO).collect(Collectors.toList());
     }
 
-    public Optional<EventModel> getEventById(int id) {
-        return eventRepository.findById(id);
+    public Optional<EventDTO> getEventById(int id) {
+        return eventRepository.findById(id).map(eventMapper::toDTO);
     }
 
-    public EventModel saveEvent(EventModel event) {
-        return eventRepository.save(event);
+    public EventDTO saveEvent(EventDTO dto) {
+        EventModel event = eventMapper.toEntity(dto);
+        EventModel saved = eventRepository.save(event);
+        return eventMapper.toDTO(saved);
     }
 
-    public void deleteEvent(int id) {
+    public String deleteEvent(int id) {
+        if (!eventRepository.existsById(id)) {
+            throw new IllegalArgumentException("Does not found");
+        }
         eventRepository.deleteById(id);
+        return "El usuario con ID" + id + "ha sido eliminado";
     }
 }
